Working FallingSand Sim

+++ src/acaer/Scene/Renderer/Renderer.h
@@ -39,11 +39,7 @@ namespace Acaer {
          * @param y             pos y
          * @param c             color of cell (rgba)
          */
-        void RenderCell(size_t x, size_t y, vColor c);
-
-
-        //void RenderChunk(WorldChunk* chunk);
-
+        void RenderCell(s32 x, s32 y, vColor c);
 
     private:
         // ** Memberse **
